공통 info model 정리
