+++ service/memo/core/src/main/kotlin/kr/co/jiniaslog/memo/domain/memo/MemoRepository.kt
@@ -1,5 +1,6 @@
 package kr.co.jiniaslog.memo.domain.memo
 
+import kr.co.jiniaslog.memo.queries.model.SimpleMemoInfo
 import kr.co.jiniaslog.shared.core.domain.Repository
 
 interface MemoRepository : Repository<Memo, MemoId> {
