--- service/memo/core/src/main/kotlin/kr/co/jiniaslog/memo/domain/memo/SimpleMemoInfo.kt
@@ -1,7 +1,0 @@
-package kr.co.jiniaslog.memo.domain.memo
-
-class SimpleMemoInfo(
-    val id: Long,
-    val title: String,
-    val content: String,
-)
