+++ service/memo/core/src/main/kotlin/kr/co/jiniaslog/memo/queries/model/MemoReferenceInfo.kt
@@ -1,0 +1,9 @@
+package kr.co.jiniaslog.memo.queries.model
+
+import kr.co.jiniaslog.memo.domain.memo.MemoId
+import kr.co.jiniaslog.memo.domain.memo.MemoTitle
+
+data class MemoReferenceInfo(
+    val id: MemoId,
+    val title: MemoTitle,
+)
