+++ service/memo/core/src/main/kotlin/kr/co/jiniaslog/memo/queries/model/SimpleMemoInfo.kt
@@ -1,0 +1,10 @@
+package kr.co.jiniaslog.memo.queries.model
+
+import kr.co.jiniaslog.memo.domain.memo.MemoId
+import kr.co.jiniaslog.memo.domain.memo.MemoTitle
+
+class SimpleMemoInfo(
+    val id: MemoId,
+    val title: MemoTitle,
+    val references: List<MemoReferenceInfo>,
+)
